+++ gitnexus/src/core/ingestion/class-extractors/generic.ts
@@ -6,6 +6,7 @@ import type {
   ClassLikeNodeLabel,
   ExtractedClassSymbol,
 } from '../class-types.js';
+import { normalizeQualifiedName, splitQualifiedName } from '../utils/qualified-name.js';
 
 const DEFAULT_SCOPE_NAME_NODE_TYPES = new Set([
   'nested_namespace_specifier',
@@ -57,20 +58,6 @@ const CLASS_LIKE_LABELS = new Set<ClassLikeNodeLabel>([
   'Enum',
   'Record',
 ]);
-
-const normalizeQualifiedName = (value: string): string =>
-  value
-    .replace(/\s+/g, '')
-    .replace(/^::/, '')
-    .replace(/::/g, '.')
-    .replace(/\\/g, '.')
-    .replace(/\.+/g, '.')
-    .replace(/^\.+|\.+$/g, '');
-
-const splitQualifiedName = (value: string): string[] => {
-  const normalized = normalizeQualifiedName(value);
-  return normalized ? normalized.split('.').filter(Boolean) : [];
-};
 
 const extractScopeSegmentsFromNode = (
   scopeNode: SyntaxNode,
